refactor: apply review hints
Refs: FILES-844
